Adding time travel - Lifting state up, again
